mcp/mcp-trello/requirements.txt
# MCP SDK
mcp>=1.1.0

# Trello API Client
py-trello>=0.20.0

# Environment Management
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
